rafs: fix build failures of test cases
Recover all test cases and make some functions as private.
