+++ rafs/src/metadata/mod.rs
@@ -307,7 +307,7 @@ impl RafsSuper {
             return self.load_v4v5(r, &sb);
         }
 
-        return Err(einval!("invalid superblock version number"));
+        Err(einval!("invalid superblock version number"))
     }
 
     /// Store RAFS bootstrap to backend storage.
@@ -717,158 +717,4 @@ pub trait RafsInode {
 /// Trait to store Rafs meta block and validate alignment.
 pub trait RafsStore {
     fn store(&self, w: &mut RafsIoWriter) -> Result<usize>;
-}
-
-#[cfg(test)]
-mod tests {
-    use crate::metadata::{add_chunk_to_bio_desc, calculate_bio_chunk_index};
-    use nydus_utils::digest::RafsDigest;
-    use std::sync::Arc;
-    use storage::device::RafsBioDesc;
-    use storage::device::{RafsBlobEntry, RafsChunkFlags, RafsChunkInfo};
-    use storage::impl_getter;
-
-    #[derive(Default, Copy, Clone)]
-    struct MockChunkInfo {
-        pub block_id: RafsDigest,
-        pub blob_index: u32,
-        pub flags: RafsChunkFlags,
-        pub compress_size: u32,
-        pub decompress_size: u32,
-        pub compress_offset: u64,
-        pub decompress_offset: u64,
-        pub file_offset: u64,
-        pub index: u32,
-        pub reserved: u32,
-    }
-
-    impl MockChunkInfo {
-        fn new() -> Self {
-            MockChunkInfo::default()
-        }
-    }
-
-    impl RafsChunkInfo for MockChunkInfo {
-        fn block_id(&self) -> &RafsDigest {
-            &self.block_id
-        }
-        fn is_compressed(&self) -> bool {
-            self.flags.contains(RafsChunkFlags::COMPRESSED)
-        }
-        fn is_hole(&self) -> bool {
-            self.flags.contains(RafsChunkFlags::HOLECHUNK)
-        }
-        impl_getter!(blob_index, blob_index, u32);
-        impl_getter!(index, index, u32);
-        impl_getter!(compress_offset, compress_offset, u64);
-        impl_getter!(compress_size, compress_size, u32);
-        impl_getter!(decompress_offset, decompress_offset, u64);
-        impl_getter!(decompress_size, decompress_size, u32);
-        impl_getter!(file_offset, file_offset, u64);
-        impl_getter!(flags, flags, RafsChunkFlags);
-    }
-
-    #[test]
-    fn test_add_chunk_to_bio_desc() {
-        let mut chunk = MockChunkInfo::new();
-        let offset = 4096;
-        let size: u64 = 1024;
-        // [offset, offset + size)
-        chunk.file_offset = offset;
-        chunk.decompress_size = size as u32;
-
-        // (offset, end, expected_chunk_start, expected_size)
-        let data = vec![
-            // Non-overlapping IO
-            (0, 0, 0, 0, false),
-            (0, offset, 0, 0, false),
-            (offset + size, 0, 0, 0, true),
-            (offset + size + 1, 0, 0, 0, true),
-            // Overlapping IO
-            (0, offset + 1, 0, 1, true),
-            (0, offset + size, 0, size, true),
-            (0, offset + size + 1, 0, size, true),
-            (0, offset + size - 1, 0, size - 1, true),
-            (offset, offset + 1, 0, 1, true),
-            (offset, offset + size, 0, size, true),
-            (offset, offset + size - 1, 0, size - 1, true),
-            (offset, offset + size + 1, 0, size, true),
-            (offset + 1, offset + 2, 1, 1, true),
-            (offset + 1, offset + size, 1, size - 1, true),
-            (offset + 1, offset + size - 1, 1, size - 2, true),
-            (offset + 1, offset + size + 1, 1, size - 1, true),
-        ];
-
-        for (offset, end, expected_chunk_start, expected_size, result) in data.iter() {
-            let mut desc = RafsBioDesc::new();
-            let res = add_chunk_to_bio_desc(
-                *offset,
-                *end,
-                Arc::new(chunk),
-                &mut desc,
-                100,
-                Arc::new(RafsBlobEntry {
-                    chunk_count: 0,
-                    readahead_offset: 0,
-                    readahead_size: 0,
-                    blob_id: String::from("blobid"),
-                    blob_index: 0,
-                    blob_cache_size: 0,
-                }),
-            );
-            assert_eq!(*result, res);
-            if !desc.bi_vec.is_empty() {
-                assert_eq!(desc.bi_vec.len(), 1);
-                let bio = &desc.bi_vec[0];
-                assert_eq!(*expected_chunk_start, bio.offset);
-                assert_eq!(*expected_size as usize, bio.size);
-            }
-        }
-    }
-
-    #[test]
-    fn test_calculate_bio_chunk_index() {
-        let (blksize, chunk_cnt) = (1024, 4);
-
-        let io_range: Vec<(u64, u64, u32, u64)> = vec![
-            (0, 1, 0, 1),
-            (0, blksize - 1, 0, 1),
-            (0, blksize, 0, 1),
-            (0, blksize + 1, 0, 2),
-            (0, blksize * chunk_cnt, 0, chunk_cnt),
-            (0, blksize * chunk_cnt + 1, 0, chunk_cnt),
-            (0, blksize * chunk_cnt - 1, 0, chunk_cnt),
-            (blksize - 1, 1, 0, 1),
-            (blksize - 1, 2, 0, 2),
-            (blksize - 1, 3, 0, 2),
-            (blksize - 1, blksize - 1, 0, 2),
-            (blksize - 1, blksize, 0, 2),
-            (blksize - 1, blksize + 1, 0, 2),
-            (blksize - 1, blksize * chunk_cnt, 0, chunk_cnt),
-            (blksize, 1, 1, 2),
-            (blksize, 2, 1, 2),
-            (blksize, blksize - 1, 1, 2),
-            (blksize, blksize + 1, 1, 3),
-            (blksize, blksize + 2, 1, 3),
-            (blksize, blksize * chunk_cnt, 1, chunk_cnt),
-            (blksize + 1, 1, 1, 2),
-            (blksize + 1, blksize - 2, 1, 2),
-            (blksize + 1, blksize - 1, 1, 2),
-            (blksize + 1, blksize, 1, 3),
-            (blksize + 1, blksize * chunk_cnt, 1, chunk_cnt),
-        ];
-
-        for (io_start, io_size, expected_start, expected_end) in io_range.iter() {
-            let (start, end) = calculate_bio_chunk_index(
-                *io_start,
-                *io_start + *io_size,
-                blksize,
-                chunk_cnt as u32,
-                false,
-            );
-
-            assert_eq!(start, *expected_start);
-            assert_eq!(end, *expected_end as u32);
-        }
-    }
 }
